+++ app/components/wallet/WalletAddDialog.js
@@ -86,7 +86,7 @@ export default class WalletAddDialog extends Component {
       <Dialog
         className={dialogClasses}
         title={intl.formatMessage(messages.title)}
-        onOverlayClick={canClose && onCancel}
+        onOverlayClick={canClose ? onCancel : null}
         active
       >
         <div className={styles.buttonsContainer}>